refactor(experiment): remove super visibility
